reducer and actions divided in two files
