+++ src/redux/reducer.js
@@ -1,30 +1,6 @@
 import {createAction, handleActions} from 'redux-actions'
 
-const SET_LOGIN_PENDING = 'modules/Login/PENDING'
-const SET_LOGIN_SUCCESS = 'modules/Login/SUCCESS'
-const SET_LOGIN_ERROR = 'modules/Login/ERROR'
-
-const loginPending = createAction(SET_LOGIN_PENDING)
-const loginSuccess = createAction(SET_LOGIN_SUCCESS)
-const loginError = createAction(SET_LOGIN_ERROR)
-
-const callLoginApi = values => {
-  return new Promise((resolve, reject) => {
-    setTimeout(() => {
-      if (values.email !== 'admin@example.com' || values.password !== 'admin') {
-        return reject(new Error('Invalid email and password'))
-      }
-      return resolve(true)
-    }, 1000)
-  })
-}
-
-export const login = values => ({
-  type: [loginPending, loginSuccess, loginError],
-  payload: {
-    data: () => callLoginApi(values)
-  }
-})
+import {SET_LOGIN_PENDING, SET_LOGIN_SUCCESS, SET_LOGIN_ERROR} from './actions'
 
 const initalState = {
   requestPending: false,
